finish task, update readme!

+++ src/services/addressService.ts
@@ -4,23 +4,15 @@ const API_BASE_URL = process.env.REACT_APP_API_BASE_URL;
 
 export const getAddresses = async () => {
   try {
-    // const response = await fetch(`${API_BASE_URL}`);
-    // if (!response) throw new Error("Failed to fetch addresses");
-    // return response.json().addresses;
-    console.log("API_BASE_URL", API_BASE_URL);
     let addressData;
     await axios
       .get(`${API_BASE_URL}`)
       .then(function (response) {
-        // const theA = response.data.addresses;
         addressData = response.data.addresses;
-        // setAddresses(theA);
-        // console.log(addressData);
       })
       .catch(function (error) {
         console.log(error);
       });
-    console.log(addressData);
     return addressData;
   } catch (error) {
     console.error("Error fetching addresses:", error);
@@ -28,10 +20,29 @@ export const getAddresses = async () => {
   }
 };
 
-export const getAddress = async (id: number) => {
-  return await fetch(`${API_BASE_URL}/address/${id}`, { method: "GET" });
+export const getAddress = async (id: string | undefined) => {
+  let addressData;
+  try {
+    await axios
+      .get(`${API_BASE_URL}/address/${id}`)
+      .then(function (response) {
+        addressData = response.data.address;
+      })
+      .catch(function (error) {
+        console.log(error);
+      });
+    return addressData;
+  } catch (error) {
+    console.error("Error fetching addresses:", error);
+    return {};
+  }
 };
 
 export const deleteAddress = async (id: number) => {
-  return await fetch(`${API_BASE_URL}/address/${id}`, { method: "DELETE" });
+  const isAddressConfirmedForDeletion = window.confirm(
+    `Do you want to Delete Address: id - ${id}?`
+  );
+  if (isAddressConfirmedForDeletion) {
+    return await fetch(`${API_BASE_URL}/address/${id}`, { method: "DELETE" });
+  }
 };
